+++ source/Private/Get-ResouceDependenciesFromMof.ps1
@@ -9,7 +9,7 @@ function Get-ResouceDependenciesFromMof
         $MofFilePath
     )
 
-    $MofFilePath = [System.IO.Path]::GetFullPath($MofFilePath)
+    $MofFilePath = Resolve-RelativePath -Path $MofFilePath
 
     $resourceDependencies = @()
     $reservedResourceNames = @('OMI_ConfigurationDocument')