+++ source/Private/Resolve-RelativePath.ps1
@@ -1,0 +1,16 @@
+function Resolve-RelativePath
+{
+    [CmdletBinding()]
+    [OutputType([String])]
+    param
+    (
+        [Parameter(Mandatory = $true)]
+        [ValidateNotNullOrEmpty()]
+        [String]
+        $Path
+    )
+
+    $currentLocation = Get-Location
+    $fullPath = [System.IO.Path]::GetFullPath($Path, $currentLocation)
+    return $fullPath
+}